Extract cvd monitor functionality into a library
Refactors the `cvd monitor` command to extract its core monitoring logic
into a reusable library (`monitor`). This allows other commands to
easily invoke log monitoring.

Bug: b/519304405
